+++ http/src/test/scala/com/twitter/finatra/http/tests/response/ResponseBuilderTest.scala
@@ -5,8 +5,8 @@ import com.twitter.finagle.http.{Cookie => FinagleCookie, Request, Response, Sta
 import com.twitter.finagle.stats.StatsReceiver
 import com.twitter.finatra.http.internal.marshalling.MessageBodyManager
 import com.twitter.finatra.http.response.ResponseBuilder
-import com.twitter.finatra.http.routing.FileResolver
 import com.twitter.finatra.json.FinatraObjectMapper
+import com.twitter.finatra.utils.FileResolver
 import com.twitter.inject.{Test, Mockito}
 import com.twitter.util.Await
 import java.io.{File, FileWriter}