+++ jackson/src/test/scala/com/twitter/finatra/json/tests/internal/streaming/StreamingTest.scala
@@ -4,6 +4,7 @@ import com.twitter.concurrent.AsyncStream
 import com.twitter.finagle.http.{Method, Request}
 import com.twitter.finatra.json.FinatraObjectMapper
 import com.twitter.finatra.json.internal.streaming.JsonStreamParser
+import com.twitter.finatra.json.tests.internal.{CaseClassWithSeqBooleans, FooClass}
 import com.twitter.inject.Test
 import com.twitter.io.Buf
 import com.twitter.util.Await
@@ -42,6 +43,47 @@ class StreamingTest extends Test {
     )
   }
 
+  test("bufs to json with some bigger objects") {
+    assertParsedSimpleObjects(
+      AsyncStream(Buf.Utf8("""[{ "id":"John" }"""),
+          Buf.Utf8(""",{ "id":"Tom" },{ "id":"An" },"""),
+          Buf.Utf8("""{ "id":"Jean" }]""")),
+      expectedInputStr = """[{ "id":"John" },{ "id":"Tom" },{ "id":"An" },{ "id":"Jean" }]""",
+      expected = AsyncStream(FooClass("John"), FooClass("Tom"), FooClass("An"), FooClass("Jean"))
+    )
+  }
+
+  test("bufs to json with some complex objects") {
+    assertParsedComplexObjects(
+      AsyncStream(Buf.Utf8("""[{ "foos":[true,false,true] }"""),
+          Buf.Utf8(""",{ "foos": [true,false,true] },"""),
+          Buf.Utf8("""{ "foos":[false] }]""")),
+      expectedInputStr = """[{ "foos":[true,false,true] },{ "foos": [true,false,true] },{ "foos":[false] }]""",
+      expected = AsyncStream(CaseClassWithSeqBooleans(Seq(true,false,true)),
+          CaseClassWithSeqBooleans(Seq(true,false,true)), CaseClassWithSeqBooleans(Seq(false)))
+    )
+  }
+
+  test("bufs to json with some bigger objects and an escape character") {
+    assertParsedSimpleObjects(
+      AsyncStream(Buf.Utf8("""[{ "id":"John" }"""),
+        Buf.Utf8(""",{ "id":"Tom" },{ "id""""),
+        Buf.Utf8(""":"An" },{ "id":"Jean" }]""")),
+      expectedInputStr = """[{ "id":"John" },{ "id":"Tom" },{ "id":"An" },{ "id":"Jean" }]""",
+      expected = AsyncStream(FooClass("John"), FooClass("Tom"), FooClass("An"), FooClass("Jean"))
+    )
+  }
+
+  test("bufs to json with some bigger objects but with json object split over 2 bufs") {
+    assertParsedSimpleObjects(
+      AsyncStream(Buf.Utf8("""[{ "id":"John" }"""),
+          Buf.Utf8(""",{ "id":"Tom" },{ "id""""),
+          Buf.Utf8(""":"An" },{ "id":"Jean" }]""")),
+      expectedInputStr = """[{ "id":"John" },{ "id":"Tom" },{ "id":"An" },{ "id":"Jean" }]""",
+      expected = AsyncStream(FooClass("John"), FooClass("Tom"), FooClass("An"), FooClass("Jean"))
+    )
+  }
+
   test("parse request") {
     val jsonStr = "[1,2]"
     val request = Request(Method.Post, "/")
@@ -53,7 +95,7 @@ class StreamingTest extends Test {
       request.writer.close()
     }
 
-    Await.result(parser.parseArray[Int](request.reader).toSeq()) should equal(Seq(1, 2))
+    assertFutureValue(parser.parseArray[Int](request.reader).toSeq(), Seq(1, 2))
   }
 
   private def readString(bufs: AsyncStream[Buf]): String = {
@@ -72,6 +114,29 @@ class StreamingTest extends Test {
 
     readString(bufs) should equal(expectedInputStr)
     val parser = new JsonStreamParser(FinatraObjectMapper.create())
-    Await.result(parser.parseArray[Int](bufs).toSeq) should equal(Await.result(expected.toSeq()))
+    assertFuture(parser.parseArray[Int](bufs).toSeq(), expected.toSeq())
   }
+
+  private def assertParsedSimpleObjects(
+    bufs: AsyncStream[Buf],
+    expectedInputStr: String,
+    expected: AsyncStream[FooClass]
+  ): Unit = {
+
+    readString(bufs) should equal(expectedInputStr)
+    val parser = new JsonStreamParser(FinatraObjectMapper.create())
+    assertFuture(parser.parseArray[FooClass](bufs).toSeq(), expected.toSeq())
+  }
+
+  private def assertParsedComplexObjects(
+    bufs: AsyncStream[Buf],
+    expectedInputStr: String,
+    expected: AsyncStream[CaseClassWithSeqBooleans]
+  ): Unit = {
+
+    readString(bufs) should equal(expectedInputStr)
+    val parser = new JsonStreamParser(FinatraObjectMapper.create())
+    assertFuture(parser.parseArray[CaseClassWithSeqBooleans](bufs).toSeq(), expected.toSeq())
+  }
+
 }